+++ index.html
@@ -229,5 +229,6 @@
 
   <script src="/js/functions.js"></script>
   <script type="module" src="/js/main.js"></script>
+  <script type="module" src="/js/draw-photos.js"></script>
 </body>
 </html>
