+++ js/util.js
@@ -1,0 +1,28 @@
+const getRandomNumber = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomObjectArray = (builder, max) => Array.from({length: getRandomNumber(0, max)}, builder);
+
+function createIdGenerator (min, max) {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomNumber(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomNumber(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+}
+
+const getRandomArrayElement = (array) => array[getRandomNumber(0, array.length - 1)];
+
+export {getRandomNumber, getRandomObjectArray, createIdGenerator, getRandomArrayElement};
